refactored ikostrikov wrappers
