+++ a2c_ppo_acktr/wrappers/data_collectors.py
@@ -1,251 +1,17 @@
-import math
-import abc
-from collections import OrderedDict
 import numpy as np
 
 from gym.spaces import Tuple
 import torch
 import gtimer as gt
 
-from a2c_ppo_acktr.model import Policy
-from a2c_ppo_acktr.algo.a2c_acktr import A2C_ACKTR
-from a2c_ppo_acktr.algo.ppo import PPO
-from a2c_ppo_acktr import utils
 from a2c_ppo_acktr.storage import RolloutStorage
 
-from rlkit.torch.torch_rl_algorithm import TorchTrainer
 from rlkit.core.external_log import LogPathCollector
-from rlkit.core.rl_algorithm import BaseRLAlgorithm
-from rlkit.data_management.replay_buffer import ReplayBuffer
 import rlkit.torch.pytorch_util as ptu
 from rlkit.util.multi_queue import MultiQueue
 import rlkit.pythonplusplus as ppp
 
 from gym_taxi.utils.spaces import Json
-
-
-class WrappedPolicy(Policy):
-    def __init__(
-        self,
-        obs_shape,
-        action_space,
-        device,
-        base=None,
-        base_kwargs=None,
-        deterministic=False,
-        dist=None,
-        num_processes=1,
-        obs_space=None,
-    ):
-        super(WrappedPolicy, self).__init__(
-            obs_shape, action_space, base, base_kwargs, dist, obs_space
-        )
-        self.deterministic = deterministic
-        self.rnn_hxs = torch.zeros(num_processes, 1)
-        self.masks = torch.ones(num_processes, 1)
-        self.device = device
-
-    def get_action(self, inputs, rnn_hxs=None, masks=None, valid_envs=None):
-        # print(inputs.shape)
-        # inputs = torch.from_numpy(inputs).float().to(self.device)
-
-        if rnn_hxs is None:
-            rnn_hxs = self.rnn_hxs
-        if masks is None:
-            masks = self.masks
-
-        value, action, action_log_probs, rnn_hxs, probs = self.act(
-            inputs, rnn_hxs, masks, self.deterministic
-        )  # Need to be careful about rnn and masks - won't work for recursive
-
-        agent_info = {
-            "value": value,
-            "probs": action_log_probs,
-            "rnn_hxs": rnn_hxs,
-            "dist": probs,
-        }
-        explored = action_log_probs < math.log(0.5)
-        # return value, action, action_log_probs, rnn_hxs
-        return (action, explored), agent_info
-
-    def reset(self):
-        pass
-
-
-class A2CTrainer(A2C_ACKTR, TorchTrainer):
-    def __init__(
-        self,
-        actor_critic,
-        value_loss_coef,
-        entropy_coef,
-        use_gae,
-        gamma,
-        gae_lambda,
-        use_proper_time_limits,
-        lr=None,
-        eps=None,
-        alpha=None,
-        max_grad_norm=None,
-        acktr=False,
-    ):
-        super(A2CTrainer, self).__init__(
-            actor_critic,
-            value_loss_coef,
-            entropy_coef,
-            lr,
-            eps,
-            alpha,
-            max_grad_norm,
-            acktr,
-        )
-        # unclear if these are actually used
-        self.eval_statistics = OrderedDict()
-        self._need_to_update_eval_statistics = True
-        self.use_gae = use_gae
-        self.gamma = gamma
-        self.gae_lambda = gae_lambda
-        self.use_proper_time_limits = use_proper_time_limits
-        self.initial_lr = lr
-
-    def decay_lr(self, epoch, num_epochs):
-        utils.update_linear_schedule(self.optimizer, epoch, num_epochs, self.initial_lr)
-
-    def train(self, batch):
-        self._num_train_steps += 1
-        self.train_from_torch(batch)
-
-    def train_from_torch(self, batch):
-
-        with torch.no_grad():
-            next_value = self.actor_critic.get_value(
-                batch.obs[-1], batch.recurrent_hidden_states[-1], batch.masks[-1]
-            ).detach()
-
-        # compute returns
-        batch.compute_returns(
-            next_value,
-            self.use_gae,
-            self.gamma,
-            self.gae_lambda,
-            self.use_proper_time_limits,
-        )
-        # update agent - return values are only diagnostic
-        value_loss, action_loss, dist_entropy = self.update(batch)
-        # TODO: add loss + entropy to eval_statistics.
-
-        # re-initialise experience buffer with current state
-        batch.after_update()
-
-    def get_diagnostics(self):
-        return self.eval_statistics
-
-    def end_epoch(self, epoch):
-        self._need_to_update_eval_statistics = True
-
-    @property
-    def networks(self):
-        return [self.actor_critic]
-
-    def get_snapshot(self):
-        return dict(actor_critic=self.actor_critic)
-
-
-# TODO: merge A2C + PPO trainers
-class PPOTrainer(PPO, TorchTrainer):
-    def __init__(
-        self,
-        actor_critic,
-        value_loss_coef,
-        entropy_coef,
-        use_gae,
-        gamma,
-        gae_lambda,
-        use_proper_time_limits,
-        lr=None,
-        eps=None,
-        clip_param=None,
-        ppo_epoch=None,
-        num_mini_batch=None,
-        max_grad_norm=None,
-        acktr=False,
-    ):
-        super(PPOTrainer, self).__init__(
-            actor_critic,
-            clip_param,
-            ppo_epoch,
-            num_mini_batch,
-            value_loss_coef,
-            entropy_coef,
-            lr,
-            eps,
-            max_grad_norm,
-        )
-        # unclear if these are actually used
-        self.eval_statistics = OrderedDict()
-        self._need_to_update_eval_statistics = True
-        self.use_gae = use_gae
-        self.gamma = gamma
-        self.gae_lambda = gae_lambda
-        self.use_proper_time_limits = use_proper_time_limits
-        self.initial_lr = lr
-
-    def decay_lr(self, epoch, num_epochs):
-        utils.update_linear_schedule(self.optimizer, epoch, num_epochs, self.initial_lr)
-
-    def train(self, batch):
-        self._num_train_steps += 1
-        self.train_from_torch(batch)
-
-    def train_from_torch(self, batch):
-
-        with torch.no_grad():
-            next_value = self.actor_critic.get_value(
-                batch.obs[-1], batch.recurrent_hidden_states[-1], batch.masks[-1]
-            ).detach()
-
-        # compute returns
-        batch.compute_returns(
-            next_value,
-            self.use_gae,
-            self.gamma,
-            self.gae_lambda,
-            self.use_proper_time_limits,
-        )
-        # update agent - return values are only diagnostic
-        value_loss, action_loss, dist_entropy = self.update(batch)
-        # TODO: add loss + entropy to eval_statistics.
-
-        # re-initialise experience buffer with current state
-        batch.after_update()
-
-        """
-        Save some statistics for eval
-        """
-        print(
-            f"value loss {value_loss}, action loss {action_loss}, dist_entropy {dist_entropy}"
-        )
-        if self._need_to_update_eval_statistics:
-            self._need_to_update_eval_statistics = False
-            """
-            Eval should set this to None.
-            This way, these statistics are only computed for one batch.
-            """
-            self.eval_statistics["Value Loss"] = value_loss
-            self.eval_statistics["Action Loss"] = action_loss
-            self.eval_statistics["Distribution Entropy"] = dist_entropy
-
-    def get_diagnostics(self):
-        return self.eval_statistics
-
-    def end_epoch(self, epoch):
-        self._need_to_update_eval_statistics = True
-
-    @property
-    def networks(self):
-        return [self.actor_critic]
-
-    def get_snapshot(self):
-        return dict(actor_critic=self.actor_critic)
 
 
 def _flatten_tuple(observation):
@@ -532,105 +298,3 @@ class HierarchicalStepCollector(RolloutStepCollector):
     def get_snapshot(self):
         return dict(env=self._env, policy=self._policy.learner)
 
-
-class IkostrikovRLAlgorithm(BaseRLAlgorithm, metaclass=abc.ABCMeta):
-    def __init__(
-        self,
-        trainer,
-        exploration_env,
-        evaluation_env,
-        exploration_data_collector: RolloutStepCollector,
-        evaluation_data_collector: RolloutStepCollector,
-        replay_buffer: ReplayBuffer,
-        batch_size,
-        max_path_length,
-        num_epochs,
-        num_eval_steps_per_epoch,
-        num_expl_steps_per_train_loop,
-        num_trains_per_train_loop,
-        use_linear_lr_decay,
-        num_train_loops_per_epoch=1,
-        min_num_steps_before_training=0,
-    ):
-        super().__init__(
-            trainer,
-            exploration_env,
-            evaluation_env,
-            exploration_data_collector,
-            evaluation_data_collector,
-            replay_buffer,
-        )
-        self.batch_size = batch_size
-        self.max_path_length = max_path_length
-        self.num_epochs = num_epochs
-        self.num_eval_steps_per_epoch = num_eval_steps_per_epoch
-        self.num_trains_per_train_loop = num_trains_per_train_loop
-        self.num_train_loops_per_epoch = num_train_loops_per_epoch
-        self.num_expl_steps_per_train_loop = num_expl_steps_per_train_loop
-        self.min_num_steps_before_training = min_num_steps_before_training
-        self.use_linear_lr_decay = use_linear_lr_decay
-
-        assert (
-            self.num_trains_per_train_loop >= self.num_expl_steps_per_train_loop
-        ), "Online training presumes num_trains_per_train_loop >= num_expl_steps_per_train_loop"
-
-    def _train(self):
-        self.training_mode(False)
-
-        for epoch in gt.timed_for(
-            range(self._start_epoch, self.num_epochs), save_itrs=True
-        ):
-            print(f"in train, with eval to go: {self.num_eval_steps_per_epoch}")
-            for step in range(self.num_eval_steps_per_epoch):
-
-                self.eval_data_collector.collect_one_step(
-                    step, self.num_eval_steps_per_epoch
-                )
-            gt.stamp("evaluation sampling")
-            print("done with eval")
-
-            for _ in range(self.num_train_loops_per_epoch):
-                # this if check could be moved inside the function
-                if self.use_linear_lr_decay:
-                    # decrease learning rate linearly
-                    self.trainer.decay_lr(epoch, self.num_epochs)
-
-                for step in range(self.num_expl_steps_per_train_loop):
-                    self.expl_data_collector.collect_one_step(
-                        step, self.num_expl_steps_per_train_loop
-                    )
-                    gt.stamp("data storing", unique=False)
-
-                rollouts = self.expl_data_collector.get_rollouts()
-                self.training_mode(True)
-                self.trainer.train(rollouts)
-                gt.stamp("training", unique=False)
-                self.training_mode(False)
-
-            self._end_epoch(epoch)
-
-    def evaluate(self):
-        self._start_epoch = 0
-        self.training_mode(False)
-
-        for epoch in gt.timed_for(
-            range(self._start_epoch, self.num_epochs), save_itrs=True
-        ):
-
-            for step in range(self.num_eval_steps_per_epoch):
-                self.eval_data_collector.collect_one_step(
-                    step, self.num_eval_steps_per_epoch
-                )
-            gt.stamp("evaluation sampling")
-
-            self._end_epoch(epoch)
-
-
-class TorchIkostrikovRLAlgorithm(IkostrikovRLAlgorithm):
-    def to(self, device):
-        for net in self.trainer.networks:
-            net.to(device)
-
-    def training_mode(self, mode):
-        for net in self.trainer.networks:
-            net.train(mode)
